+++ Assets/AccelByte/AccelBytePlugin.cs
@@ -119,6 +119,9 @@ namespace AccelByte.Api
 
             AccelBytePlugin.config = wholeJsonText.ToObject<Config>();
             AccelBytePlugin.config.Expand();
+            AccelBytePlugin.config.CheckRequiredField();
+            AccelByteSettings.Instance.UpdateConfig(AccelBytePlugin.config);
+            AccelByteSettings.Instance.Save();
             AccelBytePlugin.coroutineRunner = new CoroutineRunner();
             AccelBytePlugin.httpWorker = new UnityHttpWorker();
             ILoginSession loginSession = new LoginSession(