+++ Assets/AccelByte/Models/LobbyModels.cs
@@ -298,6 +298,7 @@ namespace AccelByte.Models
         [DataMember] public string partyAttributes;
         [DataMember] public string tempParty;           // used to store userIds to form temp party with (please include the matchmaking starter's userIds)
         [DataMember] public bool isTempParty;           // used for cancel matchmaking when using temp party
+        [DataMember] public string extraAttributes;
     }
 
     [DataContract]